requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
langchain
langchain-openai
langchain-community
faiss-cpu
pydantic
openai
python-multipart
redis
PyJWT
